new file:   build.xml ant build file. Just created for webservices. Need J2EE for wsgen and wsimport. I used the libraries provided by glassfish
	new file:   ons.solubility.data/src/ons/solubility/ws/ONService.java
the webservice. I don't want my login/password to be stored in the jar. The service
read it in a hidden file in ${HOME}

  ant publish-webservice

	new file:   ons.solubility.data/src/ons/solubility/ws/test/ONServiceTest.java
Just a client to test the webservice
Performs a few tests and echo it on the screen. The webservice must run.

 ant test-webservice

 ons.solubility.data/src/ons/solubility/data/Measurement.java

removed @XMLElement, JAXB generate it by default